edit folder structure and edit function to es6
