+++ src/redux/ducks/jokes.ts
@@ -1,14 +1,14 @@
-import {Joke, JokesState, JokesActionFetch, JokesActionAdd} from '../types'
+import { Joke, JokesState, JokesActionFetch, JokesActionAdd } from '../../types'
+import { put, takeEvery, call } from 'redux-saga/effects';
+
+import { fetchJokes as fetchJokesFromApi } from '../../api';
 
 export const ADD_JOKES = 'ADD_JOKES';
 export const FETCH_JOKES = 'FETCH_JOKES';
 
-
-
 const defaultState: JokesState = {
   jokes: [],
 };
-
 
 export const jokesReducer = (
   state: JokesState = defaultState,
@@ -22,14 +22,14 @@ export const jokesReducer = (
   }
 };
 
-export function addJokes(jokes: Joke[]): JokesActionAdd {
+export const addJokes = (jokes: Joke[]): JokesActionAdd => {
   return { type: ADD_JOKES, jokes: jokes };
 }
 
-export function fetchJokes(
+export const fetchJokes = (
   jokesLength: number,
   jokesAmount: number
-): JokesActionFetch {
+): JokesActionFetch => {
   return {
     type: FETCH_JOKES,
     props: {
@@ -37,4 +37,17 @@ export function fetchJokes(
       jokesAmount: jokesAmount,
     }
   };
+}
+
+function* fetchJokesWorker(action: JokesActionFetch) {
+  const data = yield call(
+    fetchJokesFromApi,
+    action.props.jokesLength,
+    action.props.jokesAmount
+  );
+  yield put(addJokes(data));
+}
+
+export function* jokesWatcher() {
+  yield takeEvery(FETCH_JOKES, fetchJokesWorker);
 }
