+++ src/styled/Joke.js
